What a mess.  Sorry everyone.
I'll merge these changes back in probably, but a little more
carefully.

This reverts commit 9093456bc76d95a006d224ce56fb4ce55f3c6c3f.
This reverts commit e3e0383ec34ceedff66e4e7b49501e18633c60f9.
This reverts commit d6dd9a00ee6d7e32ed2b08f7288f018ddd143029.
This reverts commit 29ba3800121c9643c49eba92f21e799e27b0cc8e.
This reverts commit 9f3d0c34ff91e600a44b8545f3752f821bcd85b9.
This reverts commit 90ed8cc581a748219a589cb173c0c38dc5284365.
This reverts commit ebefcba82bbb6362e47ee16d4822f0ee213fe4b0.

+++ src/Clients/MainApp/FSpot.Imaging/IOChannel.cs
@@ -30,8 +30,8 @@
 using System;
 using System.IO;
 using System.Runtime.InteropServices;
-
 using GLib;
+using Hyena;
 
 namespace FSpot.Imaging {
 
@@ -210,7 +210,7 @@ namespace FSpot.Imaging {
 			if (error != IntPtr.Zero)
 				throw new GException (error);
 
-			return read;
+			return (int)read;
 		}
 
 		[DllImport("libglib-2.0-0.dll")]
@@ -261,7 +261,7 @@ namespace FSpot.Imaging {
 
 		public override long Seek (long position, SeekOrigin origin)
 		{
-				  throw new NotSupportedException ();
+      		      throw new NotSupportedException ();
 		}
 
 		[DllImport("libglib-2.0-0.dll")]
