+++ src/Clients/MainApp/FSpot.Widgets/FolderTreeModel.cs
@@ -30,14 +30,18 @@
 //
 
 using System;
-
-using FSpot.Core;
-using FSpot.Database;
+using System.Collections.Generic;
 
 using Gtk;
+using GLib;
 
+using FSpot;
+using FSpot.Database;
+using FSpot.Core;
 using Hyena;
 using Hyena.Data.Sqlite;
+
+using Mono.Unix;
 
 namespace FSpot.Widgets
 {
@@ -199,7 +203,7 @@ namespace FSpot.Widgets
 					last_count += (int)GetValue (iter, 1);
 					SetValue (iter, 1, last_count);
 				}
-				count_all += last_count;
+				count_all += (int)last_count;
 			}
 		}
 	}
